Reorganized routes, created TripForm for use between CreateTrip & EditTrip. EditTrip functionality still needs to be completed
