+++ src/components/trips/TripCard.js
@@ -7,7 +7,7 @@ export default function TripCard(props) {
     const parsedEndDate = new Date(trip.endDate).toISOString().slice(0, 10);
 
     return(
-        <div className="card-container w-50 m-4 text-center rounded-lg shadow-xl">
+        <div className="card-container w-25 m-4 text-center rounded-lg shadow-xl">
             <p className='font-medium'>
                 <Link to={`/show-trip/${trip._id}`} state={{tripID: trip._id}}>{ trip.name }</Link>
             </p>
